Bug 2037247 - Part 1: Create a new parent component for the SportsWidget  a=dmeehan
Original Revision: https://phabricator.services.mozilla.com/D299065

Differential Revision: https://phabricator.services.mozilla.com/D303535
