+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/home/ui/Homepage.kt
@@ -11,7 +11,6 @@ import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
@@ -42,7 +41,6 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import mozilla.components.feature.top.sites.TopSite
-import mozilla.components.support.utils.ext.isLandscape
 import mozilla.telemetry.glean.private.NoExtras
 import org.mozilla.fenix.GleanMetrics.History
 import org.mozilla.fenix.GleanMetrics.HomeBookmarks
@@ -55,7 +53,7 @@ import org.mozilla.fenix.components.components
 import org.mozilla.fenix.compose.MessageCard
 import org.mozilla.fenix.compose.home.HomeSectionHeader
 import org.mozilla.fenix.debugsettings.sportswidget.SportsWidgetDebugTool
-import org.mozilla.fenix.ext.isLargeWindow
+import org.mozilla.fenix.ext.settings
 import org.mozilla.fenix.home.bookmarks.Bookmark
 import org.mozilla.fenix.home.bookmarks.interactor.BookmarksInteractor
 import org.mozilla.fenix.home.bookmarks.view.Bookmarks
@@ -79,10 +77,8 @@ import org.mozilla.fenix.home.recentvisits.view.RecentlyVisited
 import org.mozilla.fenix.home.sessioncontrol.CollectionInteractor
 import org.mozilla.fenix.home.sessioncontrol.MessageCardInteractor
 import org.mozilla.fenix.home.setup.ui.SetupChecklist
-import org.mozilla.fenix.home.sports.ui.CountdownPromoCard
-import org.mozilla.fenix.home.sports.ui.FollowTeamPromoCard
-import org.mozilla.fenix.home.sports.ui.MatchCard
 import org.mozilla.fenix.home.sports.ui.SportsCountrySelectorBottomSheet
+import org.mozilla.fenix.home.sports.ui.SportsWidget
 import org.mozilla.fenix.home.store.HeaderState
 import org.mozilla.fenix.home.store.HomepageState
 import org.mozilla.fenix.home.store.NimbusMessageState
@@ -220,14 +216,12 @@ internal fun Homepage(
                             }
 
                             if (sportsWidgetState.isShown) {
-                                SportsWidgetSection(
+                                SportsWidget(
                                     sportsWidgetState = sportsWidgetState,
                                     onDismiss = interactor::onSportsWidgetDismissed,
                                     onCountdownWidgetDismiss = interactor::onCountdownWidgetDismissed,
                                     onViewSchedule = interactor::onViewScheduleClicked,
-                                    onFollowTeam = {
-                                        showSportsCountrySelector = true
-                                    },
+                                    onFollowTeam = { showSportsCountrySelector = true },
                                     onSkip = interactor::onSkippedFollowTeam,
                                 )
                             }
@@ -563,50 +557,6 @@ private fun CollectionsSection(
                 )
             }
         }
-    }
-}
-
-@Composable
-private fun SportsWidgetSection(
-    sportsWidgetState: SportsWidgetState,
-    onDismiss: () -> Unit,
-    onCountdownWidgetDismiss: () -> Unit,
-    onViewSchedule: () -> Unit,
-    onFollowTeam: () -> Unit,
-    onSkip: () -> Unit,
-) {
-    Spacer(modifier = Modifier.height(44.dp))
-
-    val isLargeWindow = LocalContext.current.isLargeWindow()
-    val isLandscape = LocalContext.current.isLandscape()
-    val modifier = Modifier.fillMaxWidth(
-        fraction = when {
-            isLargeWindow || isLandscape -> 0.7f
-            else -> 1f
-        },
-    )
-
-    if (sportsWidgetState.isCountdownShown) {
-        val worldCupKickoffDate = "2026-06-11T00:00:00Z"
-        CountdownPromoCard(
-            dateInUtc = worldCupKickoffDate,
-            onViewSchedule = onViewSchedule,
-            onDismiss = onCountdownWidgetDismiss,
-            modifier = modifier.padding(horizontal = horizontalMargin),
-        )
-    } else if (sportsWidgetState.isFollowTeamsCardShown) {
-        FollowTeamPromoCard(
-            onFollowTeam = onFollowTeam,
-            onSkip = onSkip,
-            onDismiss = onDismiss,
-            modifier = modifier.padding(horizontal = horizontalMargin),
-        )
-    } else if (sportsWidgetState.matchCardState != null) {
-        MatchCard(
-            state = sportsWidgetState.matchCardState,
-            onMenuClick = {},
-            modifier = modifier.padding(horizontal = horizontalMargin),
-        )
     }
 }
